sensors/rak16000/requirements.txt
Adafruit-GPIO==1.0.3
Adafruit-PureIO==1.1.9
mock==4.0.3
pi-ina219==1.4.0
spidev==3.5
